internal/frontend: move fetchserver and versions to their own packages
Moving fetchserver to its own package will allow us to to
move the fetch logic, which isn't used by cmd/pkgsite because it
doesn't have a postgres database, out of the frontend and pkgsite. The
tests, which depend on a postgres database are moved out too, removing
a few users of the postgres database in the tests. The versions code
is moved into its own package as well so it can be used by the
fetchserver code without depending on the frontend package.

Most of server_test.go, which was testing features that are provided
by what is now the FetchServer has been moved to the fetchserver
package. To ensure that git properly realizes it as a move, the rest
of server_test has been moved to frontend_test.

In some cases I made copies of functions (absoluteTime,
insertTestModules) instead of creating intermediate packages to
contain them.

For golang/go#61399

Change-Id: Ic94419f9d75f766e289cc3fb9a1521173cefb4d1
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/519335
Reviewed-by: Robert Findley <rfindley@google.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Michael Matloob <matloob@golang.org>