internal/testing/fakedatasource: implement PostgresDB interface
This change adds the rest of the methods from the PostgresDB interface
to the fake data source so it can be used to test some functions that
expect a postgres db. Most of the methods are left unimplemented. I
removed the IsPostgresDB that I originally put on the PostgresDB
interface since now we will have other types try to be PostgresDBs. In
an upcoming CL, I will extend the DataSource interface so that it's
the same as PostgresDB, and just have the frontend handle errors that
are errors.ErrUnsupported.

The CL then updates tests on some of the methods on the frontend that need
a PostgresDB to use the FakeDataSource instead.

This CL then removes the code that sets up and runs the
internal/frontend tests with a PostgresDB.

For golang/go#61399

Change-Id: I3bae5adc5ec03536b2074eee32b4e5558dfa7478
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/520878
Run-TryBot: Michael Matloob <matloob@golang.org>
Reviewed-by: Robert Findley <rfindley@google.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>