internal/stdlib: remove dependence on internal/testing/testhelper
The testhelper package says it should only be depended on by test
files, but it's depended on by an internal/stdlib non-test file
(albeit for a testing purpose). Remove the dependency by making a copy
of the TestDataPath function. An alternative to this would be
splitting the testGoRepo into a different package, and making the goRepo
interface and the withGoRepo function exported.

For golang/go#61399

Change-Id: Ib64c18c6311f99501fc340801eee7a5625b95730
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/522242
Run-TryBot: Michael Matloob <matloob@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Robert Findley <rfindley@google.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>