internal/licences: use relative path to open testdata
We don't need to do the extra work to construct the that's being done
in testDataPath because the test will be run in the directory of the
test file. In addition, this is broken on the android builders because
they build on a different machine than where the tests run.

For golang/go#61209

Change-Id: I066cde3d972e21d52240d9c518524569d60a65f7
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/524798
kokoro-CI: kokoro <noreply+kokoro@google.com>
Run-TryBot: Michael Matloob <matloob@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Bryan Mills <bcmills@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>