internal/testing/htmlcheck: replace cascadia for htmlcheck.NotIn
This change reimplements a subset of cascadia's query functionality so
we can replace it for htmlcheck.NotIn. It adds code that parses a
subset of the css selector syntax (currently it's more than we need
for NotIn but not enough for In) and adds a query function to query
the selector in the given html node. Unlike cascadia, our query
function can match the node itself.

Future cls will add support for more selector syntax so we can support
everything htmlcheck.In is used for but we only intend so support just
what we need. I also limit the selector syntax to ascii for
simplicity.

For #61399

Change-Id: Ia03cc8a9ab42ae11d445a650e1ca0a07ee8a391f
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/541437
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>