internal/frontend: write version of ProcessReadme with rsc.io/markdown
This change adds a version of ProcessReadme that uses rsc.io/markdown
instead of goldmark. It includes most of the rewriting logic that the
goldmark version of the code has except that the heading ids are
rewritten slightly differently since we don't have access to the
original markdown which the goldmark version uses. A future CL will
probably change the processing the goldmark version does to produce
heading ids to bring them in line.

There are still many other diffs with the goldmark version of the code
but most of them are a result of rsc.io/markdown not yet supporting
GitHub-flavored markdown.

For #61399

Change-Id: Ic8981bc241448268337dc3e98a537a71146a676d
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/547915
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>