internal/fetch: handle error in root package
If a module had a problem with the package at its root, we would
end up with two packageVersionStates. That caused a DB error
in upsertPackageVersionStates.

Distinguish between the module unit and package units in a couple
of places. That addresses the root cause of the problem.

Also, fix an unrelated bug that was causing a panic when we
tried to send the error to the errorreporting service.

Change-Id: Ia02faea50f4477999c6c6eca0674b8dde0f15bb3
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/564076
Reviewed-by: Michael Matloob <matloob@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>