worker: display request state on main page
Add a RequestInfo.State field to record the current state of a request
for display.

Add a function internal.RequestState to for updating the state,
and use it for long-running requests.

Display the state on the worker main page.

Change-Id: I1a04036982aba23cf27776fac75248972a1336f3
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/568515
Reviewed-by: Michael Matloob <matloob@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>