fix(static/frontend): increase percentage of mask-image start
The exisiting percentage of where the mask-image linear-gradient
starts currently overlaps some links when text-spacing is applied
and therefore decreases the color-contrast ratio of the link making
the text less accessible. This change increases the percentage
and thus moves the gradient to be closer to the end of the container
and less likely to overlap text and links.

Fixes b/365624081

[Before screenshot](https://screenshot.googleplex.com/By9T6ZKwv6wxPue)
[After screenshot](https://screenshot.googleplex.com/Ad4J3683xC9sQdH)