fix(static/shared): add titles to social media links
Previously the social media links did not have a tool
tip that would appear on hovering over the icons to
notify the user what the purpose of the link was for.
This change adds the aria-label value as a title to
appear as a tooltip on hovering the icon.

Fixes b/#297941481