tests/screentest: retry on small diffs
If a screentest case fails but the number of differing pixels is
small, retry the the test.

This can increase reliability when a test fails but the diffs are
tiny--20 pixels or fewer. We don't know why this happens, but it
does happen occasionally.

Note that no test passes unless every pixel matches. This change
merely retries tests that are close. We've found that to help.
(After three retries, the test will fail.)

Change-Id: Ib48858197c6206cd1ea6a850c58c1335e031347e
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/633200
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>