tests/screentest: add latest x/net version to seed
A new version of golang.org/x/net was tagged, resulting in a change
in the chip from "Latest" to "Go to latest". Add this new version
to the list of seeds to capture that change, and update the goldens.

(This is a one-time fix: future tagged versions won't change the page.
The real mistake was using the then-latest version when we added x/net.)

Change-Id: I2855307ef815b970cbcef5249adfb9041d54c573
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/634080
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>