deploy: improve artifact upload
The changes in this CL improve the uploading of build artifacts
in deploy.yaml.

Astonishingly, Cloud Build does not upload artifacts if the build fails.
(The feature request is about five years old:
https://issuetracker.google.com/issues/143836671).
This makes it impossible to debug failed screentests.
The hacky solution is to cause failed build steps to succeed.
The -neverfail flag to screentest.sh accomplishes this.
It isn't used in this CL, but a subsequent CL with more debugging
support will use it.

When the build does succeed, the directory of failed screentests does
not exist, which causes artifact upload to fail. So create the
directories along with dummy files.

The wildcard syntax for Cloud Build artifact paths does not support
'**' as documented, so provide all upload paths.

Change-Id: I1c4223f4a906836000e5deac20a5b9aa920fbb97
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/634556
Commit-Queue: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
TryBot-Bypass: Jonathan Amsterdam <jba@google.com>