internal/vuln: cache DB
Put a cache in front of the vuln DB. The cache uses the modified time
of the DB to determine when it is out of date.

The cache has no maximum size, because we don't expect the vuln DB to
be too large. Currently the unzipped JSON files total under 3 MB, as
determined by listing the bucket with gsutil ls -l. Since most of it
is strings, that should be pretty close to the memory occupied by the
unmarshaled data.

Also, make other minor changes to the package that don't affect its
behavior, mostly improved doc.

Change-Id: I374fa043805fe0fa939afac80ac0ee21e71e9167
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/636515
Reviewed-by: Robert Findley <rfindley@google.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>