internal/frontend: redirect on "std/" queries
If a search query is "std/P", where P is a a stdlib package, redirect
to P without searching.

Also, update the help page on search to mention this, and to correctly
reflect the existing behavior of always doing a search for queries
without a slash.

For golang/go#71452.

Change-Id: Idf5c09ee3e81b4c6efa44c2d76470e6f496dcbcd
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/644676
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>