static/doc: promote examples into TOC link tree.
This change alters how the table of contents (TOC) is rendered to
promote the listing of examples directly into the TOC itself as well as
avoid nesting the examples under a sub-heading where they are likely not
to be seen.

A change like this assists beginners with providing immediate awareness
that the concept of examples exists. Without this change, the whole idea
of examples are easy to miss, as they are suggested essentially only in
collapsible summary tags at the bottom of the respective identifier or
scope they are attached to.

Change-Id: Ic46d20180767c25092648355af52e935a6f45f4e
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/673435
Reviewed-by: Cherry Mui <cherryyz@google.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Matt Proud <matt.proud@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>