cmd/internal/doc: add doc command in its own module
This change adds a simpler version of cmd/pkgsite, called simply 'doc',
that will be invoked by the go command to show package documentation in
the browser. This version has a couple of small changes. The first is
that it uses the AllowNoModules option added to the server to allow no
modules to match. It also makes the logging more quiet so that it's more
in line with expectations for a go tool.  Finally, it takes the path of
the page to open in the browser to open after we start listening on the
given address. The unused functionality of cmd/pkgsite is stripped out
of the doc command. The doc command is placed into its own module, in
the cmd/internal directory so we can see the set of module dependencies
in the go.mod file is limited to the smaller set of modules that
cmd/pkgsite's dependencies were limited to.

For golang/go#68106

Change-Id: Iab9db7c02044a5e08c1f6bc6894b24a400382aae
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/674437
Reviewed-by: Michael Matloob <matloob@google.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>