static/frontend: update the logo styling to not have an explicit width
If a company wants to run their own version of pkgsite, the easiest way to differentiate it from pkg.go.dev is to put their company logo (or some text, like INTERNAL) in the logo SVG next to the GO. However, doing so increases the width of the logo. In the current CSS, a wider logo means it gets shrunk down, because the width and height are constrained by css.

This CL removes the width constraint, so that the logo is allowed to be wider.

This will have a nearly imperceptible effect on the current pkg.go.dev style: the logo will be a few pixels wider. After a few pixels' width growth it becomes constrained by the height rule.

Note: This is doubly important as a CL to upstream because it's very hard to patch in this change, as it requires a change to the minified files. It's quite hard to come up with the appropriate sed command that performs a patch to the minified files.
Change-Id: I717553204888a7438a9f2c93f428bc166f095e09
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/685955
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Chressie Himpel <chressie@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Mark Freeman <markfreeman@google.com>