cmd/internal/pkgsite: improve godoc mode search in GOROOT
If a user starts go doc -http inside of GOROOT, we create three module
getters when we only need one: We create a StdlibZip module getter which
we're not going to use because we're using a local GOROOT, we create a
gopackages module getter for the module determined as corresponding to
the current directory, which is a stdlib module, and finally we create
the NewGoPackagesStdlibModuleGetter, which is the main module getter for
the standard library in that case.

The StdlibZip module getter is not used when we're using a local GOROOT,
so don't add it in that case. And if we detect that the module for a
specified directory is std or cmd for the specified local GOROOT, skip
adding it because it will be redundant with the module getter we create
using NewGoPackagesStdlibModuleGetter specifically for the standard
libarry.

Removing the redundant module getter will prevent redundant search
results returned by each of the gopackages module getters corresponding
to the standard library.

While we're here, if there's a VERSION file in GOROOT, set the version
on the module metadata we return for the standard library to the
stdlib.VersionForTag of that version, so we can report the version in
the search results.

Fixes golang/go#74459

Change-Id: If250f715b052f57f8668b02a57b4c96725be515a
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/687918
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Matloob <matloob@google.com>
Auto-Submit: Michael Matloob <matloob@golang.org>