internal/postgres: remove the exclusive lock in upsertPath
As described in golang/go#75959, the lock in upsertPath causes a
significant amount of contention during module inserts. Meanwhile, the
upsertPaths method does not acquire a lock, and does not seem to cause
deadlocks.

We still don't understand the deadlocks avoided by this lock, but it
seems that a low level of failed transactions due to deadlock may be
preferable to the contention we're seeing currently. Let's experiment
with removing the lock.

For golang/go#75959

Change-Id: I2d577b426724010e96ffc8129ae469825fdcb938
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/718720
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Ethan Lee <ethanalee@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>